+++ NEWS.md
@@ -1,3 +1,10 @@
+## Version 0.7.3 (2018-01-22) 
+ 
+ - Fix `na_replace` (and `na_explicit`) to add levels for values if
+   they do not already exist.
+ - Add tests
+ - Fix documentation
+
 ## Version 0.7.0 (2017-08-22)
 
  - Add na_explicit and na_implicit